Refactor: relocate SchusszettelComponent interface to proper API package
- Moved SchusszettelComponent interface from `api.types` to `api` to align with project conventions
- Ensured clear separation of concerns:
    - API types (DTOs) remain in `api/types`
    - Interface now resides in `business.schusszettel.api`
    - Implementation (`SchusszettelComponentImpl`) remains in `impl.business`
- Updated all imports from:
    de.bogenliga.application.business.schusszettel.api.types.SchusszettelComponent
  to:
    de.bogenliga.application.business.schusszettel.api.SchusszettelComponent
- Confirmed consistency with rest of Bogenliga layered architecture
