+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/schusszettel/api/types/SchuetzeInfoDTO.java
@@ -1,6 +1,4 @@
-package de.bogenliga.application.services.v1.schusszettel.model;
-
-import java.util.List;
+package de.bogenliga.application.business.schusszettel.api.types;
 
 /**
  * Repräsentiert einen Schützen und seine bisherige Punktzahl im laufenden Match.
