+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/schusszettel/impl/mapper/TabletSessionMapper.java
@@ -1,4 +1,4 @@
-package de.bogenliga.application.business.schusszettel.impl.business;
+package de.bogenliga.application.business.schusszettel.impl.mapper;
 
 import de.bogenliga.application.services.v1.schusszettel.model.TabletSchusszettelDTO;
